pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "livekit>=1.0.16",
    "livekit-agents>=1.2.15",
    "livekit-plugins-cartesia>=1.2.15",
    "livekit-plugins-deepgram>=1.2.15",
    "livekit-plugins-openai>=1.2.15",
    "python-dotenv>=1.1.1",
]
